+++ packages/ui/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nsui/button
 
+## 1.0.1
+
+### Patch Changes
+
+- Fixed a bug in the Button component
+- Added the .stories file for the Button component
+
 ## 1.0.0
 
 ### Major Changes